+++ src/trajdata/dataset_specific/eth_ucy_peds/eupeds_dataset.py
@@ -317,7 +317,7 @@ class EUPedsDataset(RawDataset):
         layer_names: List[str],
         cache_path: Path,
         map_cache_class: Type[SceneCache],
-        resolution: int,
+        resolution: float,
     ) -> None:
         """
         No maps in this dataset!
@@ -325,7 +325,7 @@ class EUPedsDataset(RawDataset):
         pass
 
     def cache_maps(
-        self, cache_path: Path, map_cache_class: Type[SceneCache], resolution: int
+        self, cache_path: Path, map_cache_class: Type[SceneCache], resolution: float
     ) -> None:
         """
         No maps in this dataset!
