+++ src/trajdata/dataset_specific/raw_dataset.py
@@ -80,7 +80,7 @@ class RawDataset:
         raise NotImplementedError()
 
     def cache_maps(
-        self, cache_path: Path, map_cache_class: Type[SceneCache], resolution: int
+        self, cache_path: Path, map_cache_class: Type[SceneCache], resolution: float
     ) -> None:
         """
         resolution is in pixels per meter.
